api/requirements.txt
fastapi
uvicorn
pillow
python-multipart
python-dotenv
requests
replicate